Improving the accuracy of the CNN model, adding Confusion Matrix, Training Loss Graph, Model's Predicition Testing, ModelCheckpoint and Accuracy score
The next stage of work on the project was to test the prediction of the created model, the implementation of the model prediction testing was performed. The prediction results are presented in a graphical form (table with images and their labels and prediction results). Due to the non-satifying prediction results, work was continued on increasing the actual accuracy of the model. For this purpose, the CNN model was reduced by one layer and the filters in the remaining layers were changed (respectively 32,64,32). The dense layer was changed from 128 to 64. The configuration parameters of the input data augmentation within ImageDataGenerator were changed.  Dropout class and validation_steps parameter in fit class have been dropped. The size of the image entering the shell has been increased from (100,100) to (300, 300), batch_size has been changed from 40 to 32, the number of epochs has been increased from 20 to 100. The ratio of train and test subsets has also been reduced. In the most important categories, the ratio is now 90% train, 10% test. These decisions were motivated by the fact that the dataset consists of a relatively small number of elements and therefore the learning time should be extended and the detail of data entering the model should be increased (by reducing their compression). The padding='same' attribute has been added to the Conv2D class so that no columns or rows of source images are skipped during analysis. In addition, the ModelCheckpoint class has been added so that only results from epochs where better performance was achieved than before can be saved. These steps have resulted in a decrease in training loss and an increase in the actual accuracy of the model during testing - this is calculated using the accuracy score (the prediction label created by the trained model is compared to the label assigned to the image within the ImageDataGenerator).  The accuracy of the model itself according to the logs from the fit class remained at a similar level to the previous model. In order to better understand how the model classifies test images, a confusion matrix has been implemented, the purpose of which is to graphically represent the model's prediction and understand which categories of images it has the most trouble with. To the existing accuracy graph drawing, a training loss graph drawing has also been added.