Addition of a second model using transfer learning based on the VGG-16 model, source files that contain previously developed instructions for generating predicition table, confusion matrix and accuracy score and configuration files for both models as .h5 file
In order to present the performance of the developed models, additional source files have been added, in which the already developed instructions for displaying the confusion matrix, calculating the accuracy score and prediction on a selected set of images for both trained models are executed - the prediction results are presented in graphical form (table with images and their labels and prediction results). In order to check the performance of the created model in comparison with other available solutions, a second model has been created, but not learning from scratch, but already trained for image classification, the CNN VGG16 model, and on its basis, using transfer learning, the last layer of VGG16 model is not included and instead of it a dense layer 1024 has been added and then a dense layer 6 - according to the number of output categories. VGG-16 is one of the most popular pre-trained models for image classification, presented in 2014 but used by researchers until now. I wanted to add more dense layers however, this increased the length of the calculation very significantly. The model using transfer learning uses very similar ImageDataGenerators, the differences are - it does not use input data augmentation, the parameter shuffle=true and color_mode="rgb" was added for validation_generator and train_generator. The base model uses Imagenet weights in its configuration. In the fit() method we no longer use the values obtained from the ModelCheckpoint class, as our goal is not to train the model but to teach it to the given dataset, so we add the two previously mentioned layers and obtain the final CNN model. It can be concluded from the generated confusion matrix and prediction table for each model that the prediction results of the two models are quite similar, obviously with an advantage for the model based on VGG16, since a similar result was achieved in 50 epochs and not 100 epochs as is the case for the first model. According to the generated train accuracy and validation accuracy graphs, it can be concluded that both models were characterised as underfitting - this is fully understandable due to the fact that the selected dataset was very small which directly affects the size of the dataset for model validation. The configuration files for both models have been added as a .h5 file.